third_party/WebKit/LayoutTests/platform/win/webaudio/constructor/constantsource-expected.txt

This is a testharness.js-based test.
PASS context = new OfflineAudioContext(...) did not throw an exception. 
PASS new ConstantSourceNode() threw TypeError: Failed to construct 'ConstantSourceNode': 1 argument required, but only 0 present.. 
PASS new ConstantSourceNode(1) threw TypeError: Failed to construct 'ConstantSourceNode': parameter 1 is not of type 'BaseAudioContext'.. 
PASS new ConstantSourceNode(context, 42) threw TypeError: Failed to construct 'ConstantSourceNode': parameter 2 ('options') is not an object.. 
PASS *** Invalid constructors correctly threw errors. 
PASS node = new ConstantSourceNode(context) did not throw an exception. 
PASS node instanceOf ConstantSourceNode is equal to true. 
PASS node.offset.value is equal to 1. 
PASS node.channelCount is equal to 2. 
PASS node.channelCountMode is equal to "max". 
PASS node.channelInterpretation is equal to "speakers". 
PASS *** new AnalyserNode(context) constructed node with correct attributes. 
Harness: the test ran to completion.
